Restructure the code, so the output formater has no generic logic
